Fixed delayinit argmatchers to always finish.
The delayinit coroutines needed to be marked to run until completion
(vs being abandoned when a new prompt generation begins).

But that revealed that the wake event getting recreated for each
generation made it impossible for "run until complete" coroutines to
signal when they were ready to get a new resume timeslice.

So now the wake event is global, and abandoned yields from old prompt
generations can trigger an unnecessary wake.