Fix `fromhistory` match coroutine.
Autosuggest doesn't run `fromhistory` argmatchers on a coroutine for
performance reasons.  But that means the match pipeline ends up with no
matches if a prefix was typed to match.  And `menu-complete` doesn't
regen matches.

So, when `fromhistory` declines to run in a coroutine, it needs to
signal that matches must be regenerated.