Fix #326; rl.expandtilde doesn't expand a whole command line.
Adds an optional argument to rl.expandtilde to indicate that the input
is a whole command line, to request Readline's quirky tilde expansion to
be used and expand all tildes that are preceded immediately by a space
(or at the beginning of the input).

When the second argument is false or is omitted, then the function
behaves how it was always documented to behave (though its behavior
didn't match the documentation until v1.3.36).