Rewrite path parsing routines.
Path parsing didn't handle UNC paths quite right.  For example,
`path.getname("//foo/bar")` reported `"bar"`, but that is part of the
UNC share name; it is not a file name.

That caused incorrect coloring of the input line.  E.g. `\\clink` was
colored as though an argmatcher exists, but it is not a command name, it
is a partial network address.

Now the path parsing routines should handle UNC paths correctly.  They
should also handle the `\\?\` and `\\?\UNC\` syntaxes correctly now.

This is potentially very destabilizing change.  But the unit tests were
extremely useful in debugging and validating the rewrite.