Fix #325; hangs while typing if %PATH% contains UNC paths.
When `exec.enabled` is true, the `exec.path` setting scans the %PATH%
for executable files.

When ALL of the following are true:
- `autosuggest.enabled` is enabled,
- `autosuggest.strategy` includes
`completion`,
- `exec.enable` is enabled,
- `exec.path` is enabled,
- the `%PATH%` environment variable contains a UNC path,

Then the autosuggest feature can invoke the generator for executable
file completions and the generator can scan UNC paths, which can cause
hangs while typing.

This change uses the new `os.getdrivetype()` API to skip network drives
when generating matches in the background during autosuggest (it skips
both UNC paths and SUBST-ed drives).