Fixed `oncommand` event.
The `oncommand` event needs to fire when the command word changes
between being an alias or not.  There's no need to be precise about it,
so this change relaxes it to also fire whenever the command word offset
changes, even if the word itself hasn't changed.

This is important in order for `dir` and ` dir` to be able to resolve to
different argmatchers, and be able to delay-init the non-alias one.