Fix performance problem with history timestamps.
While parsing the history file, the line buffers are not nul terminated.
But copying the timestamps accidentally expected nul termination, and
the performance suffered greatly (a potential delay at each prompt when
history timestamps are enabled).

Now it copies the correct string length (which is short and fits in the
preallocated buffer, so no memory allocations are needed).