Avoid using argmatcher color for directories.
Typing a path component that matches an argmatcher name highlights the
directory as though it's an argmatcher.

Introduce the ability to make Lua APIs that run something on a
background thread (such as getting file attributes) and then run a Lua
callback after completion.

This makes it possible to avoid mis-coloring a directory as an
argmatcher, without affecting responsiveness.