Fix #348; pasting in ConEmu duplicates the 2nd character.
This is due to yet another bug in Readline's Callback Mode.

Readline 8.0 added an optimization to the `self-insert` command, with
the intent of speeding up pasted input.  But the optimization is
incompatible with Callback Mode (it calls `rl_read_key()`, which
bypasses Callback Mode).

Setting `_rl_optimize_typeahead` to false disables the optimization,
enabling the expected behavior.

This issue was exposed by commits 39fc497 and 461faf5 in v1.3.43.