Remove "disabled by CLINK_NOAUTORUN" message.
The message used to always print to the console, even when output was
redirected.  That could be informative and useful in some contexts.  But
it also could create undesirable output in some cases (such as a tool
that automates starting new interactive shells, but where users don't
know or care about some details about how it happens).

Instead, it should only show the message when output is not redirected.

But the upcoming clink.bat performance optimization to bail early when
CLINK_NOAUTORUN is set will make the message unreachable, and there's no
way for batch scripts to check both efficiently and reliably whether
output is redirected.

So the message is simply removed now.