Fixed #361; detect CLINK_NOAUTORUN faster.
The only reliable place to detect `%CLINK_NOAUTORUN%` is inside the
Clink executable code itself.

But that incurs the performance hit of launching Clink, which can be
noticable when spawning a large number of CMD processes rapidly.

To improve the performance, the `clink.bat` script now detects autorun
even before launching Clink.  Since the batch script has to take some
shortcuts to be safe and avoid syntax errors, the Clink executable still
also checks in order to assure reliability.