Fix inject.
`setlocal`/`endlocal` must be scoped carefully otherwise they'll revert
the tagged prompt and prevent Clink from detecting when the prompt is
shown.

If that gets reported as a problem in some other scenario, then perhaps
Clink could more aggressively retag the prompt.  But for now, the
mitigation will only be in the `clink.bat` script.