Fix match prefix display.
"blah a" was fine.

"blah a\" was wrong and included "a\" when calculating the prefix
length, even though the "a\" part was not displayed.

"blah a\b" was wrong and the completion-prefix-display-length
config variable calculated the prefix incorrectly when determining how
much of the match to replace with an ellipsis.