Improve debug Lua output.
To save space, Clink strips Lua debug info from the built-in scripts.
But then when errors give a file and line number, the file name is "?"
which makes it very hard to figure out the associated file.

The luac compiler has been adjusted to use "{file.lua}" as the chunk
name for a file when the -s flag is used to strip debug info, and these
chunk names are _not_ stripped.

Now errors can refer to "{generator.lua}:123", which is much more
informative.  Internal scripts are still omitted from certain
diagnostic info lists.