Fix obscure luafunc: macro usage issue.
If a luafunc: macro modified the input line and then invoked a
completion command, the parsed line_state was inaccurate and things
malfunctioned.

Now using `rl.invokecommand()` updates the line_state so that completion
(etc) work as expected even when a macro modifies the input line and
then invokes a completion command.