Fix #406; East Asian ambiguous width character issues.
This change rewrites the logic and math for interrogating the console
font to get the expected width of East Asian ambiguous width characters
in CJK codepages.

This should be much more accurate in CJK codepages.  However, since
Windows Terminal and ConEmu (etc) do not tell the OS what font they're
using to draw the text, it is still possible for there to be mismatches
in `auto` or `font` mode of the `terminal.east_asian_ambiguous` setting.