Fixed #406; more East Asian Ambiguous width fixes.
Two changes:

The 'auto' mode now uses the font to measure the width when the terminal
host is conhost or ConsoleZ, otherwise it assumes the width is 1.

In CJK code pages, after printing the prompt Clink asks the console for
the cursor column.  This helps compensate if the console draws the
prompt width differently than Clink expects.

This is not perfect; there are still some weird edge cases.  But there
are also edge cases in PowerShell and bash and zsh and etc.  This is
about as seamless and automatic as it's going to get.