Add performance cost tracking for events.
When the `lua.debug` setting is enabled, then the `clink-diagnostics`
command (Ctrl-X,Ctrl-Z) includes additional Lua diagnostic info.

Or when `lua.debug` is off but a numeric argument is provided (for
example Alt-1,Ctrl-X,Ctrl-Z) then it includes the additional Lua
diagnostic info.

Alt-1 or Alt-2 or Alt-3 show different amounts of additional diagnostic
info.

When the `lua.debug` setting is enabled, most diagnostic info is shown
even with Alt-1.