Fix repeating background loop in auto-suggest.
When the `completion` auto-suggest strategy encountered a `fromhistory`
argument slot in an argmatcher, it accidentally kept looping
unnecessarily in the background.  Clink stayed responsive to input, but
it was doing unnecessary work in the background.

In that case it called `reset_generate_matches`, but that starts a new
match generation id, which clears any auto-suggestion and tries to make
a new suggestion, which generates matches again.  Which hits the
`fromhistory` argument slot again and calls `reset_generate_matches`...

The core of the fix is to use `setvolatile` instead (which hadn't
existed yet when the case was originally implemented).  That marks the
matches as needing to be regenerated, without actually starting a new
generation id yet, so it doesn't trigger the cycle.