Fix #452; `history delete #` doesn't work.
The "m_bank_index" member originally didn't use the bank index enum
values and was the "next bank index" instead of the "current bank
index".  That made things very confusing.  I fixed it a couple of years
ago to be the actual bank index, consistent with the rest of the
history_db component.

But I missed fixing the off-by-one math in one method.  And that caused
the history item id to end up with an incorrect bank index.

This change should fix the last remaining off-by-one math for the bank
index in the history file iterator.

Regression introduced in v1.1.43.