Fix #474; pasting is slow in Windows Terminal and ConEmu.
Windows Terminal and ConEmu handle the ^V paste operation themselves and
send the clipboard contents as though the user is typing it all in.

Readline has an optimization for that, but it didn't work properly when
using READLINE_CALLBACKS, which Clink does.  This change fixes issues in
Readline and in Clink's input peeking so that the optimization works.