Only load script from completions dirs if recognized.
If `farble` is typed and there's a `completions\farble.lua`, it should
only be loaded if the recognizer resolves `farble` to some executable
file.

Another way to think of it is if the command word would be colored in
the `color.unrecognized` color, then it shouldn't try to demand-load a
script from a completions\ directory.