Pass recognized file path when loading from a completions\ dir.
When demand-loading scripts from completions\ directories, the fully
qualified recognized path is pass as an argument.

The script can use `...` to access the argument.  For example:

local executable_file = ...

Scripts can use it to check the executable file to verify whether it's
compatible with the argmatcher (e.g. unsupported version or unsupported
name or etc), or to register different argmatchers for different fully
qualified paths (e.g. one for c:\gnu\grep.exe and another for
c:\other\grep.exe).