Fix waiting for the recognizer.
When an input line ends (e.g. pressing ENTER) and the recognizer thread
hasn't even had a chance to start processing the latest lookups, the
input line accidentally missed waiting for the recognizer.

REPRO:
1.  Have two commands in the history, without argmatchers.
2.  Press UP, ENTER, UP, UP, ENTER.
3.  Wait for the first command to finish.

EXPECTED:  Clink should wait (briefly) for the recognizer to color the
second command correctly before running it.

ACTUAL:  Due to timing race conditions, the wait logic might not realize
it needed to wait for the recognizer.  In that case the second command
got colored using the unrecognized color.

Now it waits as intended.  The wait duration has also been shortened to
only 1.25 seconds (down from 2.5 seconds).