Prototypes (disabled) for hinting how to accept auto-suggestions.
The inline version feels overly distracting to me, and is ambiguous
about what will be accepted, and what it means for it to be accepted.
For example, using Up/Down to cycle through history shows suggestions,
and the hint can look like it's saying press the Right arrow key to
accept the history entry (and execute it).

The comment row version looks imbalanced because it appears
right-justified on the line below the end of the input line, and at the
bottom of the screen (which is most of the time) it's jarring because it
scrolls up an extra line as soon as you type a letter.

A alternative hybrid could maybe be to right-justify it on the last line
of input, and only drop to the next line if there isn't room for it on
the last line.  But then it always interferes with the right-side
prompt, plus the same problems as the inline version, and it's even less
clear what's being accepted since it can be so far away from the text it
refers to.