Avoid row jitter with USE_SUGGESTION_HINT_COMMENTROW.
This makes USE_SUGGESTION_HINT_COMMENTROW always add a comment row, even
when it will be blank, so that showing the suggestion hint doesn't shift
the display up an extra row.

But in exchange, every prompt at the bottom of the screen shifts the
display up an extra row.  The shift is briefly visible when pressing
ENTER repeatedly.