Improvements to `clink.autoupdate` and the updater.
Changed the `clink.autoupdate` setting to allow `off`, `check`,
`prompt`, and `auto` for more control over how and when updating
happens.

The setting was bool before, but now it's an enum, so a little
migration glue was needed as well.

The default is `check`, so that there is no change in behavior unless a
user opts in to new behavior by setting it to `prompt` or `auto`.