Add comments about algorithm for completion prefix coloring.
Match display filtering can return any arbitrary string, with any escape
codes anywhere.  That makes it difficult to apply completion prefix
coloring when match display filtering is active.

However, there are common patterns that make it possible to do a little
extra processing to find the match text in the display string.

And while testing the refactored match display filtering code for
`clink-select-complete`, I got confused into thinking there was a bug in
the completion prefix coloring algorithm.  But after spending enough
time in the debugger it became clear it was working as intended, and my
test case was an outlier.

So this captures the design intent to save me (and others) time in the
future.