Add WC_NO_BEST_FIT_CHARS.
Only allow exact conversions when converting ACP.  Better to convert
a character in a filename to `?` than to some similar character that
accidentally references a different file, and is undetectable by the
caller.

Also the clipboard copy passed an inflated size limit by mistake.  In
practice it's unlikely to cause a problem, but in a race condition it
might.  E.g. if the buffer gets allocated to the right size, then a
longer string is set into the clipboard by some process, and the new
string overruns the allocated buffer.