Fix `match.max_rows`.
It should control the number of rows of items, not the overall number of
rows used.

I.e. the comment row and/or showing descriptions below can use
additional rows but they should be outside the max_rows limit since
they're not necessarily present and thus have an inconsistent and
unpredictable effect on the number of rows of items actually shown.