Fix #511; `!` doesn't need to be treated as a word break.
In the CMD grammar, the `!` is a word break character.  Completion in
CMD does not automatically quote the `!` character in a filename,
because it's a word break character.

But, the two cases where it's a word break are pretty obscure:
1.  In the `set /a` command.
2.  When delayed variable expansion is enabled.

It doesn't make sense to try to use completion in `set /a` when using
the `!` operator, so it's ok for Clink to not treat `!` as a word break
character in that case.

Clink doesn't offer environment variable completion for the `!VAR!`
delayed variable expansion syntax, so it doesn't cause any new problems
if Clink doesn't treat `!` as a word break character in that case.

Running CMD interactively with delayed variable expansion enabled seems
unusual, no one has requested support for `!VAR!` completion, and there
are some tricky (maybe intractable) problems with supporting `!VAR!`
completion anyway.

So, it seems safe to remove `!` from the list of word break characters.