Report "up-to-date" as a success, not a failure.
The `clink update` command was designed as a command to run after being
notified that an update is available, so it considered "up to date" to
be a failure.  But it's very reasonable to run `clink update` to mean
"check if there's an update, and if so then install it".  So, being
up-to-date shouldn't be considered as a failure.