Add module name when invoking Expand-Archive.
Someone hit a case where something seems to be interfering with the
built-in Expand-Archive cmdlet.  An internet search found a similar
problem someone encountered, and their solution was to add the module
name when invoking the cmdlet.

That doesn't seem to cause any problems here, and maybe it will help
prevent problems for some users?