Fixed called `clink.refilterprompt()` in a coroutine.
It has to instead schedule a refilterprompt to happen when control
returns to the main coroutine so that the current working directory is
accurate.  Coroutines track cwd independently, and a long-running
coroutine could end up filtering the prompt using an out of date
context, causing inaccurate content to appear in the prompt.