Fix assertions in test program after failures.
Since clatch's REQUIRE() throws, it bypasses the normal end_line() call.
The simplest way to compensate is to make line_editor_impl's dtor call
end_line() if it hasn't been called yet.

It would be nice to catch the exception and call end_line(), but it
isn't (and shouldn't be) a public method.

Since begin_line() asserts if end_line() wasn't called, in practice this
shouldn't weaken the validations.