Added an `onparse` callback in argument positions.
This callback enables an argmatcher to conditionally skip or repeat an
argument position.

Examples:

In the FOR command the /R and /F flags accept optional arguments.  An
argmatcher can use the `onparse` callback to skip the argument position
if the next word needs to be handled by a different argument position.

In the FOR command the `(set)` syntax accepts any number of items in the
set, surrounded by parentheses.  An argmatcher can use the `onparse`
callback to repeatedly use the same argument position for multiple words
in a row.

Example:  `for /r [dir] %v in (a* b* c*) do echo %v`