Detect light or dark console themes when possible.
ConPty currently doesn't return the active color table.  But when using
legacy conhost (or if/when ConPty gets fixed) then it's possible to
detect whether the console is currently using a light or dark theme.

However, this is still undocumented for now, since Windows Terminal and
all ConPty-based terminals aren't able to detect it.