Default color.suggestion adapts to the theme.
With legacy conhost, Clink auto-detects the background color and
dynamically chooses a default suggestion color that is faint but has
sufficient contrast.

With Windows Terminal and other ConPty-based terminals there's
currently no way to auto-detect the background color, so Clink has to
use a simple default which may have too much contrast depending on the
current theme.

There's only so much Clink can do to auto-adapt to custom color themes.
If you set a custom color theme in a console window, then you should
expect to also have to customize Clink colors to fit with it.