Fix race condition in async lua tasks.
This is an obscure theoretical race condition found through code
inspection.

If an async_lua_task is running and the current lua_State is destroyed
and a new lua_State is created, a race conditon may cause malfunctions.
The existing task might not get canceled, its callback ref may become
invalid, and it might attempt to run against the new lua_State.

Now the async_lua_task list is cleared when the lua_State is destroyed.