Prototype for enumerating file shares on a server.
The intent was to enable completing share names.  But the WNet APIs
don't seem to work well for this unless they're used to traverse the
global hierarchy.  And even when passing the seemingly same inputs as
when traversing the hierachy, I still can't get them to work.

Next step would be to try the NetShareEnum API instead.