The "set" keyword is optional in the default_inputrc file.
This is an alternative approach to fixing the default_inputrc syntax
errors in older versions of Clink.  The idea is to allow the
default_inputrc files to be read successfully, even if then don't get
updated with fixes for the syntax errors.  Since overwriting an existing
default_inputrc file could potentially result in data loss if the user
had modified it.