Work around OS escape code limitation.
See #546.  When "Wrap text output on resize" is unchecked (or when
Window Width != Screen Buffer Width) then escape codes for cursor
positioning and insert/delete characters and clearing text and etc only
affect the part of the terminal display that's within the visible area
(the CONSOLE_SCREEN_BUFFER_INFO::srWindow area).

That results in jitter and some text getting displayed inaccurately.

There's only so much Clink can do to compensate for the problem, but
this change tries to mitigate it as much as possible.  However, the
visible area jitters around.  There's nothing Clink can do about that;
the behavior is built into the OS console APIs.