Fix cloning line_state_lua; forgot to copy m_shift.
That was showing up as problems during input line coloring.

Repro 1; `sudo sudo dir/` hit a script error due to out of range word
index in the lua_word_classifications, but that was because the
break_slash didn't copy m_shift when creating the new line_state_lua, so
it effectively reset the shift, but only in the line_state_lua.  So the
loop consumed the second `sudo` twice by mistake, and the
lua_word_classifications shift hadn't gotten reset, so it effectively
had too few words.

Repro 2; `start i . clink/` colored `clink` as unrecognized.  Same root
cause as repro 1.