Make rl_message prompts show the full prompt.
Readline shows only the last line of the prompt when rl_message() is
used.  But that leaves the user with no context after "complete" or etc
if the last line of the prompt is just ">", for example.

Also, the NSEARCH and READSTR prompts are not normal rl_message()
prompts, so the display code needs to treat them as normal prompts
rather than rl_message() prompts.