Fix Readline and Clink bugs when using ^C or ^Break.
When READLINE_CALLBACKS is used, Readline is missing a call to
readline_internal_teardown() when the Readline input editor ends.
That leads to multiple problems, including a double-free of
rl_undo_list.

Clink had some issues as well:
- The on_signal() module callbacks were not called when injected, and
  were only called in the standalone Lua interpreter.
- The rl_module::on_signal() should call rl_callback_handler_remove() so
  the handler gets removed at the right time.

This also fixes the long-standing quirk where the `*` modmark was
displayed briefly in the new prompt after using ^C or ^Break while a
modmark was active.