Work around WT issue in ReadConsoleOutputCharacterW.
See issue #574.  An API in Windows Terminal behaves differently than in
legacy conhost.

1. It fails, but should succeed.
2. It fails, but returns that it succeeded.
3. It claims to have succeeded, but returns that the length was 0.

This detects the failure based on the 0 length and treats it as having
failed, despite the API claiming to have succeeded.

So, at least Clink and Lua scripts can be aware that it failed, instead
of thinking it succeeded.

Since Clink's Lua API strips trailing spaces, callers have no way to
deduce failure, since the returned string is empty, which is consistent
with a line that was full of spaces.