Include debug info for the embedded scripts.
It doesn't add much to the exe sizes, and it makes a big difference for
troubleshooting when issues are reported.

The luac compiler has been modified to add a `-R` flag which ultimately
enables Clink to tell which symbols come from embedded scripts, so that
it can suppress them from the `clink-diagnostics` command.