Fix #597; prompt shows leftover text sometimes.
Regression introduced in v1.6.10.  The optimization to reduce flicker
missed the fact that multi-line prompts often don't explicitly clear
past the ends of lines in the prompt string.

This inserts a "clear to end of line" escape code before each CRLF or LF
that ends a line in a multi-line prompt string.