Fix pushing a lua_bindable created by make_new.
enumshares_lua had never actually needed to use push, because its
reference was created on the stack.  But the changes in push that were
made for it didn't actually work.

find_files_lua will need push to work properly for make_new instances,
because its :files() function will need to push a reference to it.