Work around scoop.
Scoop tries to control app versions and updates.  But it makes
assumptions which are inaccurate for some apps (such as Clink).

Fine.  Clink will try to work around that by getting the path that was
used to invoke Clink, rather than the path where Clink actually exists.
This ends up getting scoop's "current" directory symlink for the app,
instead of the real "version" directory of the app.

It looks like this shouldn't cause any regressions.

Related to issue #615.