Refactor argmatcher parser.
The motivation is to restructure how it iterates through the words, to
(greatly) simplify adding an `onalias` callback function that can expand
arbitrary words into more/other words while parsing the input line.

But this is also just a much better approach because it encapsulates the
"next word" logic, instead of having it distributed across multiple
blocks of special case logic.

This is also a step towards possibly factoring all of the parser loops
(which use "next word" to iterate) into a single routine, with callbacks
or etc to allow differences in behavior.