Revert part of 679de941dc.
The `min()` makes no sense, and caused `clink-select-complete` to
truncate descriptions e.g. for `clink set exec.`.

I'm not sure this is correct yet.  The corresponding code in
display_matches.cpp was changed in that commit, and selectcomplete_impl
needs to behave differently because it musn't allow line wrapping.

But this improves the observed issue, and I haven't yet been able to
find any cases that it breaks.