Avoid red flicker from the recognizer.
The recognizer discarded its cache when starting a new input line.  This
was a simple way to constrain growth of the cache.

But if "foo" is recognized in one input line, and is typed again a few
seconds later in the next input line, then "foo" would briefly flash red
(i.e. the unrecognized color) because the cache got discarded.

I got tired of seeing the red flicker.

Now upon starting a new input line it ages out entries to keep 2 hours
of results in the cache, but it marks all entries as out of date.  This
enables the recognizer to report the cached result, but also enqueue the
word again for recognition.