Argmatchers can provide hints.
Arguments and flags may include a `hint` field which can be a string or
a function.  If a function is given, then the function is called to
provide a hint when the cursor is in that argument slot (or flags).

The drawback is that currently it always adds extra computation even
when there are no hints.