Avoid jitter from input hints at bottom.
When the input line is at the bottom of the screen, showing the comment
row scrolls the display up one line.  This can be disorienting if it
happens often.  To mitigate that, if an input hint has been shown before
in this session then force showing the comment row even if it's blank.
This forces reserving space for a comment row once an input hint has
been shown at least once.