Call `onarg` functions during classify, gethint, etc.
As far as I can tell, the reason I restricted it to only while
generating matches was because onarg functions can change the current
directory, and I thought that was only needed while generating matches.
It seems to have been intended as a performance optimization, but it
creates functional drawbacks.

It's needed more generally, e.g. so hinters can count how many arguments
are present.  An arg_index is the index into arg slots, and it can
differ from the number of arguments provided.