Fix when `onarg` runs relative to `classify_word`.
A recent change runs onarg during classify and gethint, which exposed
the fact that onarg is run _after_ classify for flags, but _before_
classify for args.  The order should always be classify then onarg, and
if classify needs similar processing as in onarg then it should run the
logic itself instead of relying on onarg.  This is necessary to resolve
an ordering issue; otherwise if onarg changes the current directory then
classify would evaluate the word _after_ the current directory was
changed, ruining for example how i.lua colors the directory argument
based on whether it's a valid directory.