Add support for .clinkprompt files.
These are self-contained custom prompt modules, and only one at a time
can be active.

This makes it easy to have a library of many custom prompts and switch
between them quickly and easily.

Once a prompt module is loaded, it stays loaded for the rest of the
session.  But it only receives prompt filter and event calls while it's
the active prompt module.