Force built-in themes directories.
Some users remove the Clink program directory and/or Clink profile
directory from the list of scripts directories.

For `themes\` directories, Clink will forcibly look under the Clink
program directory and profile directory, so that the built-in custom
prompts and color themes and saved color themes can be found even when a
user has removed those from the list of script directories.