The display manager optimization accommodates emoji sequences.
The optimization minimizes display updates by finding the left and right
edges of the text that has changed.  But emoji sequences need to be
considered as a unit, otherwise the optimization can split a sequence
and accidentally fail to update the display accurately.

For the left index, the accommodation is simple.

For the right index, the accommodation would have been complicated
because parsing emoji sequences in reverse is ambiguous.  So instead,
now it snaps the right index to the nearest ASCII character boundary.