More accurately predict width of variant selectors applied to non-emoji.
Caveats:

Windows Terminal 1.22 is greatly improved wrt emoji rendering, but it
still has some quirky behavior.  In particular, it gets very confused
when a skin tone variant selector follows a line drawing character if
"Built-in Glyphs" is enabled.

The Unicode spec describes how it imagines a skin tone variant selector
"by itself" should be rendered.  But "by itself" is ambiguous.  Windows
Terminal applies a skin tone variant selector to whatever glyph precedes
it.  So to show a skin tone chit you have to precede the selector with
some other character, such as a space.  But it even applies skin tone
selectors to letters, and superimposes a skin tone chit over the
preceding letter.  It's not unreasonable behavior, but at least on the
surface it appears to depart from my interpretation of the Unicode spec.
But it's also possible that Windows Terminal is intentionally matching
behavior of some Linux or iOS terminal for compatibility purposes.