Fix #718; error about orphaned coroutine.
When autosuggest.strategy lists `complete` before `history` (or if
`history` isn't listed), then it's possible for an edge case timing
condition to occur which can result in a "coroutine is orphaned" error.

The fix is to discard a coroutine that doesn't pass the generation
check, regardless whether it has a yieldguard.  But that means also
`release_coroutine_yieldguard` must release a yieldguard that doesn't
pass the generation check, regardless whether the yieldguard is ready.

Commit 6dea3d630b91750f8cdc533211cd1540c993f952 introduced this issue
while fixing a different issue, because it didn't use quite the right
approach.