Fix #718; error about orphaned coroutine.
When autosuggest.strategy lists `complete` before `history` (or if
`history` isn't listed), then it's possible for an edge case timing
condition to occur which can result in a "coroutine is orphaned" error.

The fix is to discard a coroutine that doesn't pass the generation
check, once its yieldguard is ready.  This avoids the error, while still
preventing a bunch of `popenyield` calls from running concurrently (and
eating system resources).

Commit 6dea3d630b91750f8cdc533211cd1540c993f952 introduced this issue
while fixing a different issue, because it didn't use quite the right
approach.