Fix #727; crash when any error occurs in `io.popen()`
The C runtime has a bug in its `_popen()` function, which is called by
Lua's `io.popen()` function.  The C runtime attempts to close a file
handle twice.  The C runtime behavior for invalid parameters defaults to
using the safest default behavior, which is to terminate the process.

Lua and Clink are both compatible with (and expect) the previous
standard C runtime behavior, which was to gracefully return an error to
the caller and let the caller choose how to respond to the error.

This change configures the C runtime to gracefully return errors to the
caller instead of terminating the cmd.exe process.

This also updates the oh-my-posh.clinkprompt to produce friendlier
error messages when that happens.