Prototype code (DISABLED) for []{} completions.
Issue #731 observed that Clink behaves slightly differently than CMD
when completing a word containing []{} characters without any quoting.
That's because of the match.substring setting.  Turning that off is the
only way to match CMD behavior.

Then the request flipped to instead ask for behaving even less like CMD,
to have a way to not treat []{} characters (and some other characters)
as word breaks during completion.  This prototype change allows
exploring the consequences if the input line parser word breaks are
changed in that way.

It would affect the parser in general and could affect any scripts that
use line_state.