Remove prototype code related to #731.
The prototype revealed that the approach creates other problems.

With enough complexity and testing and time investment, maybe some kind
of solution could be designed and implemented.  But the benefits are so
niche and the complexity and cost so great that it is not something to
which I will commit any more of my own time.