Fixed #738; Windows Terminal mishandles `CSI K` past end of line.
When the printed text fills a screen line, word wrapping will happen.
But the cursor does not wrap to the next line until _additional_ text is
printed that's important so escape codes and CR (0x0D, Carriage Return)
can be used at the end of a screen line without wrapping.

But the cursor exists at a "phantom" position.  The screen is only X
cells wide, but the cursor is at a phantom position of X+1.

When Windows Terminal receives the `CSI K` code (`ESC [ K` aka EL or
Erase in Line) while the cursor is in that phantom position, Windows
Terminal accidentally clears the last character on the screen line and
resets the cursor back to position X instead of letting it stay at X+1.

That results in both incorrect display and also incorrect cursor
position.  Since the cursor is no longer at the phantom position, the
next text that's printed does not wrap, and instead moves the cursor to
the phantom position again.  So the output gets garbled.

Fortunately, Clink already keeps track of the cursor position, and can
avoid printing the Erase in Line code in cases where Windows Terminal
will mishandle it.

NOTE:  This cannot fix *all* possible places where Clink or Lua scripts
may use the Erase in Line code.  But it's able to work around the type
of case reported in issue #738.

A fix is still needed in Windows Terminal in order for the Erase in Line
code to work correctly in Windows Terminal in general (e.g. when used by
programs other than Clink).