Fix long-standing mystery bug!
If the current line has a modmark and the clink-popup-history command is
used to choose a history line to execute, then the modified current line
gets updated into the history list and the modmark is cleared.

The clink-flex-prompt project has an `oninputlinechanged` callback which
calls `clink.refilterprompt()` if the modmark state changes for a line.

The `oninputlinechanged` callback is called _after_ the transient prompt
has been displayed.  So, at that point, Clink and Readline are done with
the current line -- no further display updates should be allowed.

Safely and accurately blocking further display updates is a tricky
proposition, but I think this change accomplishes it.

CAVEAT:  If a script directly prints text during callbacks, then Clink
has to let it.  This change only prevents Clink from trying to redisplay
the prompt and/or input line at a script's request, if Clink knows it's
not supposed to be do that anymore (until the next input line).

THE RISK:  There might be places that print stray bits of text without
checking whether it's ok to print them.  The display wouldn't have shown
up correctly anyway, though, so this shouldn't create new problems.