Fix undo list management in search modes.
NOTE:  When using noninc search, Readline does not move the history
index.  Instead it inserts the found item into the current input line...
Or into the current history line, if the current history index is not at
the bottom of the history list.  That seems like a bug, but it exists in
bash, so it's not unique to Clink.