Fix heap corruption in `os.enumshares()`.
The iter_aux for enumshares_lua calls push() to push a new reference to
the object for the iterator to use later.

The push() calls bind() (as it must), and since the object was not yet
in the Lua registry, bind() created a new uservalue object.  So then two
different Lua object instances pointed at the same enumshares_lua
instance.

make_new() needs to always register the uservalue object, so that push()
can work without cross-linking the object.